Add property to remove JsonNullable import
The openApiNullable property controls the use of JsonNullable in the templates, and by consequence in generated classes. Since in this generator we are using our own templates that make no use of the JsonNullable wrapper but the property is not set, we are getting an extra import that is unused but breaks compilation if you don't have the needed dependency in your project. This commit fixes this by hardcoding this property to false.