test: add test of JSON request bodies
Add a test that fails due to the request body types (here: PatchRequestDto) getting generated with @QueryParam annotations on each field. This results in PATCH (or POST) requests that have each field as a query param as part of the URL ("/patch?name=Max&age=42") instead of as part of a JSON body.

This currently fails only if run against RESTEasy Reactive Client. The legacy RESTEasy Classic Client generates the expected body.

To my knowledge the unwanted behavior of RESTEasy Reactive is however to spec and the @QueryParam annotation should not be added to types used as RequestBody. Of course this means that one should never share types used both for query request parameters and request bodies.