Introduce support for truly non-blocking APIs with resteasy-reactive and mutiny (#387)
* feat(api): return response instead of void

Imagine POSTing to a resource that creates something and returns no
payload but a 201 with a Location header. Right now, there is no way to
get that url. After this commit, it
will be possible

* feat(api): support non-blocking apis

Introduce a ew config property `mutiny`, which will change the Api
classes to return `Uni<X>` instead of `X` for every method. This is
required to make truly non-blocking api calls with
quarkus-resteasy-reactive.

Of course, this requires `quarkus-resteasy-reactive` on the classpath.

* Update deployment/src/main/java/io/quarkiverse/openapi/generator/deployment/SpecItemConfig.java

Co-authored-by: Helber Belmiro <helber.belmiro@gmail.com>

* Apply suggestions from code review

Co-authored-by: Helber Belmiro <helber.belmiro@gmail.com>

* Update deployment/src/main/resources/templates/api.qute

Co-authored-by: Helber Belmiro <helber.belmiro@gmail.com>

---------

Co-authored-by: Helber Belmiro <helber.belmiro@gmail.com>